This adds a new implementation in the `v3.cpp` source file.

It's hella faster (30-50%), uses less memory, and is overall a cleaner
implementation!

This implementation is largely the same algorithm, with one exception:
constant values are stored in-line in the tape, as the second word
of a 64-bit instruction.

Other changes include:
- Breaking libfive opcodes into GPU-specific opcodes, which bake the
  lhs/rhs state (from a slot vs immediate) into the opcode.  For example,
  OP_SUB maps to either GPU_OP_SUB_LHS_RHS, GPU_OP_SUB_LHS_IMM, or
  GPU_OP_SUB_IMM_RHS depending on its arguments.
- Tuning all of the kernels to hit 100% occupancy (less than 32
  registers used in every kernel).

+++ inc/gpu_affine.hpp
@@ -251,7 +251,7 @@ __device__ inline Affine operator/(const float& a, const Affine& b) {
     return Affine(a) * reciprocal(b);
 }
 
-__device__ inline Affine min(const Affine& a, const Affine& b, uint8_t& choice) {
+__device__ inline Affine min(const Affine& a, const Affine& b, int& choice) {
     const Interval ia = a.as_interval();
     const Interval ib = b.as_interval();
     if (ia.upper() < ib.lower()) {
@@ -297,7 +297,7 @@ __device__ inline Affine min(const Affine& a, const Affine& b, uint8_t& choice) 
     return Affine(min(ia, ib, choice));
 }
 
-__device__ inline Affine min(const Affine& a, const float& b, uint8_t& choice) {
+__device__ inline Affine min(const Affine& a, const float& b, int& choice) {
     const Interval ia = a.as_interval();
     if (ia.upper() < b) {
         choice = 1;
@@ -309,7 +309,7 @@ __device__ inline Affine min(const Affine& a, const float& b, uint8_t& choice) {
     return Affine(ia);
 }
 
-__device__ inline Affine min(const float& a, const Affine& b, uint8_t& choice) {
+__device__ inline Affine min(const float& a, const Affine& b, int& choice) {
     const Interval ib = b.as_interval();
     if (a < ib.lower()) {
         choice = 1;
@@ -321,7 +321,7 @@ __device__ inline Affine min(const float& a, const Affine& b, uint8_t& choice) {
     return Affine(ib);
 }
 
-__device__ inline Affine max(const Affine& a, const Affine& b, uint8_t& choice) {
+__device__ inline Affine max(const Affine& a, const Affine& b, int& choice) {
     const Interval ia = a.as_interval();
     const Interval ib = b.as_interval();
     if (ia.lower() > ib.upper()) {
@@ -367,7 +367,7 @@ __device__ inline Affine max(const Affine& a, const Affine& b, uint8_t& choice) 
     return Affine(max(ia, ib, choice));
 }
 
-__device__ inline Affine max(const Affine& a, const float& b, uint8_t& choice) {
+__device__ inline Affine max(const Affine& a, const float& b, int& choice) {
     const Interval ia = a.as_interval();
     if (ia.lower() > b) {
         choice = 1;
@@ -379,7 +379,7 @@ __device__ inline Affine max(const Affine& a, const float& b, uint8_t& choice) {
     return Affine(ia);
 }
 
-__device__ inline Affine max(const float& a, const Affine& b, uint8_t& choice) {
+__device__ inline Affine max(const float& a, const Affine& b, int& choice) {
     const Interval ib = b.as_interval();
     if (a > ib.upper()) {
         choice = 1;
